requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
gradio==4.26.0
pydantic==2.5.0
python-dotenv==1.0.0
requests==2.31.0
pyngrok==5.2.3
httpx==0.25.2
aiohttp==3.9.1
